add caching

+++ watch/cli/coco_align_geotiffs.py
@@ -78,7 +78,6 @@ import shapely
 from shapely import ops
 from os.path import join, exists
 
-
 class CocoAlignGeotiffConfig(scfg.Config):
     """
     Create a dataset of aligned temporal sequences around objects of interest
@@ -135,6 +134,15 @@ class CocoAlignGeotiffConfig(scfg.Config):
             '''
             if True, normalize and draw image / annotation sequences when
             extracting.
+            '''
+        )),
+
+        'keep': scfg.Value('none', help=ub.paragraph(
+            '''
+            Level of detail to overwrite existing data at, since this is slow.
+            "none": overwrite all, including existing images
+            "img": only add new images
+            "roi": only add new ROIs
             '''
         )),
 
@@ -227,6 +235,9 @@ def main(**kw):
 
     output_bundle_dpath = dst_dpath
 
+    from pympler.tracker import SummaryTracker
+    tracker = SummaryTracker()
+
     if regions == 'annots':
         pass
     elif exists(regions):
@@ -234,6 +245,8 @@ def main(**kw):
         print('region_df = {!r}'.format(region_df))
     else:
         raise KeyError(regions)
+
+    tracker.print_diff()
 
     # Load the dataset and extract geotiff metadata from each image.
     dset = kwcoco.CocoDataset.coerce(src_fpath)
@@ -267,17 +280,19 @@ def main(**kw):
     new_dset.dataset['info'] = [
         process_info,
     ]
-
+    tracker.print_diff()
     to_extract = cube.query_image_overlaps2(region_df)
 
+    tracker.print_diff()
     for image_overlaps in ub.ProgIter(to_extract, desc='extract ROI videos', verbose=3):
+        tracker.print_diff()
         video_name = image_overlaps['video_name']
         print('video_name = {!r}'.format(video_name))
 
         sub_bundle_dpath = join(extract_dpath, video_name)
         print('sub_bundle_dpath = {!r}'.format(sub_bundle_dpath))
 
-        cube.extract_overlaps(image_overlaps, extract_dpath,
+        new_dset = cube.extract_overlaps(image_overlaps, extract_dpath,
                               rpc_align_method=rpc_align_method,
                               new_dset=new_dset, visualize=visualize,
                               write_subsets=write_subsets,
@@ -614,6 +629,8 @@ class SimpleDataCube(object):
             >>> cube, region_df = SimpleDataCube.demo(with_region=True)
             >>> to_extract = cube.query_image_overlaps2(region_df)
         """
+        import gc
+        gc.collect()
         # New maybe faster and safer way of finding overlaps?
         ridx_to_gidsx = geopandas_pairwise_overlaps(region_df, cube.img_geos_df)
         print('ridx_to_gidsx = {}'.format(ub.repr2(ridx_to_gidsx, nl=1)))
@@ -685,12 +702,13 @@ class SimpleDataCube(object):
                         'properties': region_props,
                     }
                     to_extract.append(image_overlaps)
+        gc.collect()
         return to_extract
 
     def extract_overlaps(cube, image_overlaps, extract_dpath,
                          rpc_align_method='orthorectify', new_dset=None,
                          write_subsets=True, visualize=True, max_workers=0,
-                         aux_workers=0):
+                         aux_workers=0, keep='none'):
         """
         Given a region of interest, extract an aligned temporal sequence
         of data to a specified directory.
@@ -717,6 +735,11 @@ class SimpleDataCube(object):
             visualize (bool, default=True):
                 if True, dump image and annotation visalizations parallel to
                 the extracted data.
+
+            keep (str): Level of detail to overwrite existing data at, since this is slow.
+                "none": overwrite all, including existing images
+                "img": only add new images
+                "roi": only add new ROIs
 
         Returns:
             kwcoco.CocoDataset: the given or new dataset that was modified
@@ -745,7 +768,17 @@ class SimpleDataCube(object):
         video_name = image_overlaps['video_name']
         video_props = image_overlaps['properties']
 
+        if new_dset is None:
+            new_dset = kwcoco.CocoDataset()
+
         sub_bundle_dpath = ub.ensuredir((extract_dpath, video_name))
+
+        if exists(join(sub_bundle_dpath,
+                       'subdata.kwcoco.json')) and keep == 'roi':
+            print('ROI found on disk; adding')
+            sub_dset = kwcoco.CocoDataset(
+                join(sub_bundle_dpath, 'subdata.kwcoco.json'))
+            return new_dset.union(sub_dset)
 
         latmin, lonmin, latmax, lonmax = space_box.data[0]
         dates = sorted(date_to_gids)
@@ -755,8 +788,6 @@ class SimpleDataCube(object):
             'properties': video_props,
         }
 
-        if new_dset is None:
-            new_dset = kwcoco.CocoDataset()
         new_vidid = new_dset.add_video(**new_video)
 
         for cat in dset.cats.values():
@@ -798,7 +829,7 @@ class SimpleDataCube(object):
                                   date, num, frame_index, new_vidid,
                                   rpc_align_method, sub_bundle_dpath,
                                   space_str, space_region, space_box,
-                                  start_gid, start_aid, aux_workers)
+                                  start_gid, start_aid, aux_workers, (keep == 'img'))
                 start_gid = start_gid + 1
                 start_aid = start_aid + len(anns)
                 frame_index = frame_index + 1
@@ -861,7 +892,7 @@ class SimpleDataCube(object):
 def extract_image_job(img, anns, bundle_dpath, date, num, frame_index,
                       new_vidid, rpc_align_method, sub_bundle_dpath, space_str,
                       space_region, space_box, start_gid, start_aid,
-                      aux_workers=0):
+                      aux_workers=0, keep=False):
     """
     Threaded worker function for :func:`SimpleDataCube.extract_overlaps`.
     """
@@ -883,8 +914,13 @@ def extract_image_job(img, anns, bundle_dpath, date, num, frame_index,
     objs.extend(auxiliary)
 
     is_rpcs = [obj['geotiff_metadata']['is_rpc'] for obj in objs]
-    assert ub.allsame(is_rpcs)
-    is_rpc = ub.peek(is_rpcs)
+    if not ub.allsame(is_rpcs):
+        # TODO fix this, probably WV from smart-stac and smart-imagery mixed?
+        print(objs)
+        print(is_rpcs)
+        is_rpc = False
+    else:
+        is_rpc = ub.peek(is_rpcs)
 
     if is_rpc and rpc_align_method != 'affine_warp':
         align_method = rpc_align_method
@@ -909,7 +945,7 @@ def extract_image_job(img, anns, bundle_dpath, date, num, frame_index,
         job = executor.submit(
             _aligncrop, obj, bundle_dpath, name, sensor_coarse,
             dst_dpath, space_region, space_box, align_method,
-            is_multi_image)
+            is_multi_image, keep)
         job_list.append(job)
 
     dst_list = []
@@ -1069,14 +1105,6 @@ def _write_ann_visualizations(new_dset, new_img, new_anns, sub_bundle_dpath):
 
     for chan in components:
         spec = chan.channels.spec
-        canvas = chan.finalize()
-
-        # canvas = kwimage.imread(dst_gpath)
-        canvas = normalize_intensity(canvas)
-        if len(canvas.shape) > 2 and canvas.shape[2] > 4:
-            # hack for wv
-            canvas = canvas[..., 0]
-        canvas = kwimage.ensure_float01(canvas)
 
         view_img_dpath = ub.ensuredir(
             (sub_bundle_dpath, sensor_coarse,
@@ -1087,12 +1115,23 @@ def _write_ann_visualizations(new_dset, new_img, new_anns, sub_bundle_dpath):
              '_view_ann_' + align_method))
 
         view_img_fpath = ub.augpath(name, dpath=view_img_dpath) + '_' + str(spec) + '.view_img.jpg'
-        kwimage.imwrite(view_img_fpath, kwimage.ensure_uint255(canvas))
-
-        dets = kwimage.Detections.from_coco_annots(new_anns, dset=new_dset)
         view_ann_fpath = ub.augpath(name, dpath=view_ann_dpath) + '_' + str(spec) + '.view_ann.jpg'
-        ann_canvas = dets.draw_on(canvas)
-        kwimage.imwrite(view_ann_fpath, kwimage.ensure_uint255(ann_canvas))
+
+        if not exists(view_img_fpath) and not exists(view_ann_fpath):
+            canvas = chan.finalize()
+
+            # canvas = kwimage.imread(dst_gpath)
+            canvas = normalize_intensity(canvas)
+            if len(canvas.shape) > 2 and canvas.shape[2] > 4:
+                # hack for wv
+                canvas = canvas[..., 0]
+            canvas = kwimage.ensure_float01(canvas)
+
+            kwimage.imwrite(view_img_fpath, kwimage.ensure_uint255(canvas))
+
+            dets = kwimage.Detections.from_coco_annots(new_anns, dset=new_dset)
+            ann_canvas = dets.draw_on(canvas)
+            kwimage.imwrite(view_ann_fpath, kwimage.ensure_uint255(ann_canvas))
 
 
 def update_coco_geotiff_metadata(dset, serializable=True, max_workers=0):
@@ -1476,7 +1515,7 @@ def _fix_geojson_poly(geo):
 
 
 def _aligncrop(obj, bundle_dpath, name, sensor_coarse, dst_dpath, space_region,
-               space_box, align_method, is_multi_image):
+               space_box, align_method, is_multi_image, keep):
     # NOTE: https://github.com/dwtkns/gdal-cheat-sheet
     latmin, lonmin, latmax, lonmax = space_box.data[0]
 
@@ -1531,6 +1570,8 @@ def _aligncrop(obj, bundle_dpath, name, sensor_coarse, dst_dpath, space_region,
             template = ub.paragraph(
                 '''
                 gdalwarp
+                --config GDAL_CACHEMAX 500 -wm 500
+                --debug off 
                 -te {xmin} {ymin} {xmax} {ymax}
                 -te_srs epsg:4326
                 -t_srs epsg:4326
@@ -1547,6 +1588,8 @@ def _aligncrop(obj, bundle_dpath, name, sensor_coarse, dst_dpath, space_region,
             template = ub.paragraph(
                 '''
                 gdalwarp
+                --config GDAL_CACHEMAX 500 -wm 500
+                --debug off 
                 -te {xmin} {ymin} {xmax} {ymax}
                 -te_srs epsg:4326
                 -t_srs epsg:4326
@@ -1566,10 +1609,11 @@ def _aligncrop(obj, bundle_dpath, name, sensor_coarse, dst_dpath, space_region,
             dem_fpath=dem_fpath,
             SRC=src_gpath, DST=dst_gpath,
         )
-        cmd_info = ub.cmd(command, verbose=0)  # NOQA
     elif align_method == 'affine_warp':
         template = (
             'gdalwarp '
+            '--config GDAL_CACHEMAX 500 -wm 500 '
+            '--debug off '
             '-te {xmin} {ymin} {xmax} {ymax} '
             '-te_srs epsg:4326 '
             '-overwrite '
@@ -1584,9 +1628,12 @@ def _aligncrop(obj, bundle_dpath, name, sensor_coarse, dst_dpath, space_region,
             xmax=lonmax,
             SRC=src_gpath, DST=dst_gpath,
         )
-        cmd_info = ub.cmd(command, verbose=0)  # NOQA
     else:
         raise KeyError(align_method)
+
+    if not (exists(dst_gpath) and keep):
+        cmd_info = ub.cmd(command, verbose=0)  # NOQA
+        # cmd_info = ub.cmd(command, verbose=1)  # NOQA
 
     return dst
